requirements.txt
websockets>=12.0
ccxt>=4.0.0
google-genai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
hmmlearn>=0.3.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
aiosqlite>=0.19.0
scikit-learn>=1.3.0
scipy>=1.11.0
